fix: critical and high Go audit findings in PSPF builder and launcher
Critical fixes:
- Handle all 6 ignored Seek errors in builder.go with proper error checking
- Replace magic number 1 with io.SeekCurrent in all Seek calls
- Fix IntegritySeal.Algorithm from "ecdsa-p256" to "ed25519"
- Fix SOURCE_DATE_EPOCH parsing to use strconv.ParseInt for Unix timestamps

High fixes:
- Split IsProcessRunning into platform-specific files with build tags
  (locking_unix.go uses Signal(0), locking_windows.go uses OpenProcess)
- Replace hardcoded ~/.cache logic with workenv.GetCacheRoot() in
  execution.go and execution_env.go for cross-platform consistency
- Standardize ExecutionInfo JSON tag from "env" to "environment" to match
  builder_types.go ExecutionConfig
