+++ src/flavor-go/pkg/psp/format_2025/execution_env.go
@@ -8,10 +8,7 @@ import (
 	"strings"
 
 	"github.com/hashicorp/go-hclog"
-)
-
-const (
-	defaultCacheSubdir = ".cache/flavor/workenv"
+	"github.com/provide-io/flavor/go/flavor/internal/workenv"
 )
 
 // setFlavorCacheBeforeWorkenv sets FLAVOR_CACHE to the HOST's cache directory.
@@ -24,14 +21,8 @@ func setFlavorCacheBeforeWorkenv(env []string, logger hclog.Logger) []string {
 		return env
 	}
 
-	// Get HOME from parent environment BEFORE it gets overwritten by workenv
-	homeValue := getenv(env, "HOME", "")
-	if homeValue == "" {
-		logger.Warn("⚠️ HOME not found in environment, skipping FLAVOR_CACHE setup")
-		return env
-	}
-
-	flavorCache := filepath.Join(homeValue, defaultCacheSubdir)
+	// Use workenv.GetCacheRoot() for cross-platform cache directory consistency
+	flavorCache := filepath.Join(workenv.GetCacheRoot(), "workenv")
 	env = append(env, fmt.Sprintf("FLAVOR_CACHE=%s", flavorCache))
 	logger.Debug("🗂️ Setting FLAVOR_CACHE to HOST cache", "path", flavorCache)
 	return env
